Fixed an old bug
This was introduced in an old code cleaning commit. Since this was
always a NOP, this looked like a bug. Now after analysing it, and deemed it
as a bug, it is finally fixed. The hiding code is now executed instead
of just skipping it and this fixes the old regression.