Removing globals 12
The .UI_disabled global variable is now removed completely and
replaced with (semi)global reactive value.

This fixes the crash, if the user clears environment together with the
hidden values.

Removing globals made the responsibilities clearer, but the current
behaviour is incorrect. THe 1 second ad hoc waiting causes race
conditions that break the user expectations and leads to incorrect
states leading to mysterious error messages.

The proper way to fix this is to disable all the tabs and eliminate the
waiting and the semiglobal reactive value altogether.