Fixing UI disable/enable 1
This commit introduces the new way of disabling and enabling the UI
elements. When launching a lengthy operation in Test & Submit tab panel,
the Exercises tab panel is properly disabled and enabled after the
operation has finished. After extending this to Login tab panel, this
can ne taken to be the behavior in all the tab panels.

After this the UI_disabled can be removed.