Cleaning observer implementation 6
The login tab panel now uses only observeEvent to follow the changes.
This is more robust, more performant and cleaner. In addition, now the
login button is not activated when username (and server) are missing.

The credential handling has still a type error, i.e. the credentials
should be named list with entries provided during initialisation. Now
the implementation is trusting that $ operator returns NULL for
nonexisting named elements.