Fixing disabling/enabling 1-8
This commit introduces a more flexible UI enabling / disabling method
that is reactive.

The submit tab panel reacts whether the user is logged in and if an
exercise is selected. The `UI_state` and various sublists in `UI_elements`
are used to obtain a reactive state for the enabling / disabling. The
input elements and data is followed via observers to update the
`UI_state`. The course tab uses multiple reactive values to overcome
Shiny reactions, but this is still not optimized.

The goal is to do as little and observe as little, and keep and fix
all the UI glitches there are.