Submission and progress bar 1
The commit fixes a big that could crash the addin if the connection
happened to break after the submission reached the server and before
server responded.

The progress bar styling was accidentally overriden and caused an UI
bug. This commit changes that and makes the progress bar bit more
eye friednly. The inlined CSS should be separated for theming.